cpu/sh4: add a few patches to fool AIRTRIX.
Add a few debug points and fix logging of 'sync' while we are at it.

The patches all work around the currently missing emulation of the AICA
and IO boards. With these AIRTRIX reaches the 'WARNING' screen, writes
texture data into TEXRAM and is desperately trying to write text on
screen; but the GPU code doesn't know how to that yet. I'm working on it
right now.