hikaru/gpu: add proof-of-concept 1A040000 emulation.
The device is used to copy texture rectangles around in TEXRAM. In
particular, AIRTRIX uses it to write ASCII text on the framebuffer.
The result is not yet visible on the GL screen; that will be
implemented shortly. The text however is clearly visible in TEXRAM with
any viewer for raw graphics.