hikaru/gpu: update the CP insns format and description table.
According to the Demul authors, CP opcodes are only 9-bit wide (as opposed to
the 11-bits we were using previously). This doesn't make much of a difference,
emulation-wise, but it's cleaner.

We take this opportunity to factorize instruction execution; a bunch of
renderer-related cleanups will follow.

Also, more updated docs (a couple of bits were, again, donated by the
Demul authors.)