hikaru/{gpu,renderer}: simplify and clean up viewport handling.
Attempts to properly set up the viewport and depth test. Uses all available
gpu_viewport_t info (except for the queue thing, which probably does some
advanced stuff -- such as enabling depth buffer writes -- that we don't care
about for the moment).

Now the perspective is fine in both the BOOTROM and BRAVEFF.