hikaru: make SGNASCAR work.
The ROMBD config was wrong. This makes SGNASCAR not attempt to read from
beyond the ROM boundaries, and consequently work up to the intro.

However, at some point the game writes bogus values to CMDRAM and the CP
interpreter hits an assertion. It looks like a CPU/CP synchronization error.