vk/state: add anew.
It implements a very dumb directo-to-FILE interface.

The real way to do savestating is serialization and keeping copies of the
current machine in memory and if anything goes wrong discard the copy loaded
from the save state etc. That's one long sentence, and it tells the truth.

However, i) this is a C program, and ii) let's be pragmatic.