vk/machine: fix save-states, they work with BRAVEFF now.
The problem was that the first loop loaded the SH4 IREGs from the save
state, while the second one reset the two SH4, and thus the IREGs. This
lead to BRAVEFF waiting forever on a DMA operation that couldn't run
because of DMAOR=0...

(BTW, this also resolves to a rather large slow-down in e.g. PHARRIER.)